chore(functions): add lib/lti.js module
Refactor LTI logic out of the server handler code.
